Changes which implies some use of the `Precursor' keyword to simplify the complexity of inheritance.
Since we can now set the background color of a windows in WEL, I've adpated this
feature to EiffelVision. This implies a new implementation of `class_background', with
a synonym `background_brush' and they are defined once in WIDGET_IMP. This avoids
duplication of the same code.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@10410