1. The parent and the child or the children of a widget (`parent_imp', `child', `child1', `child2') are no longer stored in the interface class, but in the implementation-interface class (_I).
2. A new attribute (interface) int ev_widget_i allow us to have an acces to the interface of a widget implementation (it is used in gtk and to give the parent of a widget to the user - not done yet).

3. the feature add_child is no more visible for the user, because it is now located in ev_container_i.

4. Some calculus have been removed from the windows implementation. Because before, when a child was
added to a container, we used to notify the parent that the child had changed its size. It is no more
usefull, because now, when we create a widget, it directly notify the parent of its size, its minimum parameters and things like that.

All the following classes have been adapted to this change.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@11466