Previous new implementation of WEL_RICH_EDIT_STREAM was buggy when calling `ceif_wean (cookie)'. Because `cookie' was an integer and `ceif_wean' took a ANY as argument it was doing a metamorphosis of `cookie' into an INTEGER_REF and was trying to free the INTEGER_REF, whereas we wanted to free the pointer contained in `cookie'.
As a result, we reviewed the implementation of the new mechanism using `cookie' and
now `cookie' are no more exported because it could imply to many troubles if someone
else was playing with it at the GC level. The current implementation should be safe
now.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@16564