`copy_from_real_dialog' now removes the old implementation's menu bar, and sets it into `Current'. This fixes bug with disappearing menu's when you did the following:
create a dialog with a menu bar/menus.
Show relative
show
show_relative

At one of these stages the menu bar would be missing, as it was not
correctly moved from the old implementation, meaning it could not be set
correctly in `Current'.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@32829