Fixed bug in `background_pixmap' and `set_background_pixmap'. The problem manifested if you performed the following:
internal_pixmap.set_with_named_file ("C:\icon_cmd_history_color.ico")
l_fixed_1.set_background_pixmap (internal_pixmap)
internal_pixmap.set_with_named_file ("C:\Jazz Man.png")
l_fixed_2.set_background_pixmap (internal_pixmap)

The pixmap displayed on the first fixed (in this case) would have the
second image. This is because we did not clone the implementation of
the pixmap when we set it into a container, but cloned it when
querying the pixmap. The fix was to reverse this, and perform the clone
when the pixmap was added, and not when it was queried.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@33559