Added `docked_out_actions' and `docked_in_actions'.
Changed signature of `add_external' so that the window to which the dialog
will be relative must be specified.

Fixed both `restore_heights_post_insertion' and `restore_minimized_tool'
so that if the size of `Current' has been reduced, they will not force
`Current' to enlarge.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@37384