New agent implementation from the compiler implies new library classes for ROUTINE, FUNCTION and PROCEDURE.
With this implementation we went from about a 5s execution time to a 2s
execution time on an integration test made by Peter Haefliger using agents to
evaluate a mathematical function. The improvement was mostly due to the fact
that we now have just one TUPLE `internal_operands' which holds both closed
and open operands at their correct position (to avoid reinserting closed arguments
to their correct position). Of course `open' operands are not set. When we perform
the call, we take the given tuple for the open operands and assign its
values to their proper position in `internal_operands'. Then the call
is done through using `internal_operands'. The C generated code in `ececil.c'
now accepts a TUPLE object as first argument and simply extract the values
from the TUPLE and perform the call.

After a call is done through `call' or `item' we need to clean up the
reference items of the open operands otherwise the GC will still track
them. However since most of the time we do not need to do that, we have
a boolean flag `is_cleanup_needed' to tell if we should do it or not.
The non reference items keep their value. Doing this improves perf by about 0.3s.

We use special feature `$$_result_type' from `inline C specific' which
enables us to get the result type of the external and perform an efficient
call to `rout' in the case of FUNCTION. Otherwise we would have to do it
like we were doing it before (through a C function from our runtime) which
was not efficient at all (we gain about .5s by doing this).

Finally because we do not use a C intermediate struct to put values passed
to the Eiffel routine, we do not need to stop the GC and therefore it fixes
a bug in a multithreaded application where the GC will be disabled if 2
concurrent calls are performed at the same time. This also improves the
general performance of agents.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@37444