Redefined `set_foreground_color' to invalidate both controls comprising `Current' which is required now that the foreground and background color may be set on a combo box.
Redefined `set_focus' to fix a bug which manifested when a combo box had
the focus, lost the focus and then was destroyed. If the window containing
the combo box was then set the focus again, the window would not raise to
the front.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@38028