Major Improvement in flickering of toolbars and surrounding widgets, when a property of a tool bar button is changed. Previously, the implementation of `reset_button' called `remove_button' followed by `insert_button' which caused resizing to occur twice, moving any widgets displayed next to the tool bar if necessary and then moving them back. Now the resizing does not occur until the end, and the update of the window is locked. This fix grealty improves the visible flickering of tool bars.
To reproduce:
Create two toolbars in a horizontal box.
Add a button in the first, and a number of buttons to the second.
keep modifying the pixmap of the first, and you will see the second flicker
as it moves left and then right again.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@41328