Fixed bug in update of list items when adding a pixmap that caused the text of the item to get cut off.
To reproduce, do the following.

create a list and display on screen wide enough to display complete text and
image.
create a list item with a text.
Insert item into list.
Set a pixmap to the item -> Now the text is cut off.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@42482