Fixed the following bug:
Start a pick and drop transport somewhere within the system.
Drop onto a widget that has a `pebble' for a pick and drop transport.
Call `process_events' on the application from within the drop actions.

This causes the pick to start from the target which is incorrect.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@47161