Fixed resizing issue which prevented part of the grid to be redrawn when it should have done. This could be reproduced in EiffelStudio as follows:
Target a class such as STRING which has many features into the editor.
Autocomplete a feature with no typed text.
scroll to the very bottom of the list.
Close the autcomplete.
Perform another autocomplete.
Now resize the window to have a greater height, by dragging down on the bottom
of the window.

The newly exposed area is displayed as grey. This is because we were not always
keeping our last vertical and horizontal scroll bar positions up to date with
all change to the scroll bar values.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@53406