Fixed bug in the handling of the default push button. We were calling `enable_default_push_button' on the default push button. However, this did not work in the case where a focus in event was received on another button during the showing of the dialog. We were ending up in a state where two buttons thought they were the both the default push button. To fix this, we now call `set_default_push_button' which handles the unsetting of the previous button so we do not end up in the state where two buttons thought they were the default at once.
To reproduce this:
Turn on "check", "require" and "ensure" and recompile EiffelStudio.
Launch EiffelStudio and from the startup dialog, select "new project"
Click "OK" on the new project settings dialog and as the question dialog
is popped up, a check fails in the EiffelVision2 implementation.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@54469