Fixed bug in grid which caused rows to be hidden/drawn over each other as reported by Jocelyn and found in the debugger. To reproduce: 	grid.enable_tree 	add_items (1, 2) 	grid.row (1).insert_subrows (6, 1) 	grid.pointer_double_press_item_actions.extend (agent update_row) 	... 	update_row (an_x, a_y, a_button: INTEGER; an_item: EV_GRID_ITEM) is 			-- 		local 			row: EV_GRID_ROW 		do 			row := grid.row (1) 	grid.remove_rows (row.index + 1, row.index + row.subrow_count_recursive) 		end
You must expand and collapse the first row before double clicking on the
grid. There are two rows left in the grid, but you will see that they
appear to be drawn on top of one another and chaning the selection gives
an indication of this.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@54836