Implemented persistency of TEST_SESSION_RECORD objects by storing them into the EIFGENs folder. When a project is loaded, old records are retreived and displayed in the testing tool (see test results from previous sessions).
There is a fixed number of records before the testing tool automatically starts deleting older ones. The user is able to mark records as persistent, meaning they will not automatically be deleted (needs to be properly implemented on GUI side).

Changed some semantics of TEST_SUITE_S and its implementer on when a record is added to the repository, this however will not affect the UI.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@80712