Redone the code generation of manifest array so that it doesn't yield an assertion failure in void-safe mode when creating an array with an attached actual argument type. So instead of calling `make' from ARRAY and then filling it, we create a SPECIAL instance, fill it, and then use the new `to_array' feature from SPECIAL to create the ARRAY. The changes required updating the code generation for melted code, C code and .NET code. Fixed a bug in the creation of a SPECIAL of BIT types in melted mode as we miss reading the BIT size. Added requirement for `to_array' to be present in SPECIAL. Removed code generation of precondition checking for SPECIAL creation in the case of manifest arrays as the compiler knows we are fullfilling the precondition. Modified `updated_info' from CREATE_TYPE to always return a CREATE_TYPE instance as this is what is happening in practice.
While testing the .NET code generation above, we found that `generating_type' was not working properly so we fixed it but it will not work when applied to a non-Eiffel class (we raise an exception while waiting the fix).
Fixed code generation of `default' from TYPE the same way it was done in classic to avoid a check violation when it is called. 

SVN: https://svn.eiffel.com/eiffelstudio/trunk@81426