Implemented dynamic loading for API `SHCreateMemStream'
Because prior to Windows Vista, this function was not included in the public Shlwapi.h file, nor was it exported by name from Shlwapi.dll. To use it on earlier systems, you must call it directly from the Shlwapi.dll file as ordinal 12.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@86945