Refactored error categories in ABEL.
The error hierarchy is now the same as in the following document:
https://docs.google.com/document/d/1pB-ScX9iH9yXm4M8sIUL-MnmFZsBHl38zipO4ug6I9I/edit#

Also added a constant short tag and a meaningful description to every
error cagegory. Replaced occurrences of GENERAL_ERROR and
UNRESOLVABLE_TRANSACTION_CONFLICT with a fitting error in the new
hierarchy.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@93011