Added a plug-in mechanism to the new backend.
The goal of the new mechanism is to inject external code
just before objects get written to the database, or during 
retrieval before being built by the object-relational mapping
layer.
With the old backend interface this was already possible, but 
required a lot of glue code and some dirty hacks around contracts
(usually resolved by disabling the contract). With the new system it
should be easier now to write plugins, without the need to forgo
tight contracts.
Possible use cases is the ESCHER plugin which adds versioning
information to an object, or a filter plugin to remove transient
attributes prior to storing them. 

SVN: https://svn.eiffel.com/eiffelstudio/trunk@93060