Replaced the TRANSACTION_ISOLATION_LEVEL with ANOMALY_SETTINGS.
Instead of understanding the transaction levels defined in the SQL
standard, the user can now say which anomalies he wants to allow
and which should be prevented. ABEL will then pick a suitable transaction
isolation level, based on the capabilities of the backend.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@93383