Implemented support for inserting and querying value types, such as STRING.
The changes mainly affect the deferred class VALUE_TYPE_HANDLER, which 
can now act as a normal handler if it realizes that the root object is 
a value type. The value gets wrapped in a BACKEND_OBJECT instance with
the fake attribute "ps_value_type_item", which was introduced in the
last commit.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@93565