Several changes:
*) Changed the backend interface to accept two ARRAYED_LISTs with
object primary keys and types, instead of one list filled with TUPLE
with the same information. This avoids having to create many little tuples.

*) Changed the in-memory backend to directly implement the backend interface
instead of relying on the converter utility.

*) Fixed a bug and added a regression test: When the object initialization depth
is not set to infinity, there might be cases where an object wasn't loaded to its
full depth. This was mainly because of caching inside the READ_MANAGER. The fix
is a rather drastic approach: the cache inside the read manager just gets wiped out
when the object initialization depth isn't set to `infinity'.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@93774