Bug fix.
Bug description:
Copy and paste texts with multiple lines with some leading tabs into the editor.
The cursor in the end is positioned N characters before the place it used to be.
(N is the number of lines of the pasted text)

The bug was caused by the fact that `text' read from clipboard  does not contain
%R anymore, which lead to wrong number of inserted characters counting. (Missing
the number of %R)

review#5154506417373184

SVN: https://svn.eiffel.com/eiffelstudio/trunk@93779