Some cleanup and smaller changes:
*) Enabled function inlining in ECF.
*) Changed the READ_MAANGER algorithm to always use integer intervals.
*) Added a function `object_item' to access fully initialized objects.
This is a prepraration for the next step where I want to recycle the
reflector objects and directly store initialized objects in the array.

Also added a new utility PS_LOOKUP_TABLE which is basically a
double hash table. So far it is unused however.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@93783