Adjusted traversal algorithm to use PS_IDENTIFIER_TABLE.
This new table should be faster because it can do a hash lookup
based on POINTER, instead of traversing the list. The drawback
is that garbage collection has to be disabled for a short time, 
and that the lookup table has to be recreated for every traversal.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@93841