Fix minimum size computation of windows with Aero Glass
When using a later version of the Microsoft C compiler (version 2012 or later), it uses version 6.0 which triggers Windows assuming our application is Aero Glass. The end result is that if a window contains a widget of minimum height of 25, our computation would yield a height of 15 for that widget, having the effect of being truncated. The solution is to actually query the client rect and the window rect to Window and deduce from there the extra width and height required to size the window so that it can accommodate our widget.

See about the Microsoft linker flag about C compiler and versioning
https://msdn.microsoft.com/en-us/library/fcc1zstk(v=vs.110).aspx

See The following discussion about GetWindowRect computation which helped in resolving the issue.
https://social.msdn.microsoft.com/Forums/windowsdesktop/en-US/6c1c67a9-5548-4e9b-989f-c7dbac0b1375/getwindowrect-on-nonresizable-windows-under-aero-glass

SVN: https://svn.eiffel.com/eiffelstudio/trunk@98372