Fix non-working scrollable area on GTK3
The reason was that we have to redefine gtk_insert_i_th to insert the child in the parent box, we cannot use the inherited one from EV_VIEWPORT_IMP has the parent is not a box but a layout. On GTK2 it seems to have worked by chance but not in GTK3.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@99026