Change setting of drawing mode and add non-aliasing for lines
Change the way we set the drawing mode, we now use a function to convert the Vision mode to the Cairo mode.
Ensure that when we draw lines they are pixel perfect (i.e. no extra rounded or squared extra pixel at the end of lines).

SVN: https://svn.eiffel.com/eiffelstudio/trunk@99078