Fix bug introduced at rev#99195
The effect of the bug was to not show a hidden tab when we selected it from the drop down menu. It was because removing the call to `force_extend` we were calling `on_label_selected` with the number of items when selecting the item instead of the item position at the time of creation.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@99606