Workaround to bypass an annoying issue with docking on gtk3, when the window is resized.
When the related window is resized, the split area of the docking layout are sometime frozen. Using this hack allow to move the splitter.

TODO: find better and cleaner solution, as the current commit is a dirty workaround, applied only for gtk3 implementation.

SVN: https://svn.eiffel.com/eiffelstudio/trunk@105876